Update documentation to clarify "a busy thing"
It turns out that using a configuration object, you can set the `busy` property to a boolean to start/stop the indicator, but the documentation wasn't clear on this. 

This clarifies that as an option, and provides an example of using an `async` piped boolean value into the `busy` property.